backend/requirements.txt
fastapi
uvicorn
langchain
huggingface_hub
python-dotenv
google-api-python-client
langchain-community
